Consolidate index middleware implementations
