+++ apps/api/src/routes/views.ts
@@ -5,7 +5,7 @@ import { db } from '../lib/db.js';
 import { describe, validator } from '../lib/openapi/describe.js';
 import { rejectTraceIndex } from '../middleware/reject-trace-index.js';
 import { requireUser } from '../middleware/require-user.js';
-import { withIndexMeta } from '../middleware/with-index-meta.js';
+import { withIndexMeta } from '../middleware/with-index.js';
 import { SavedViewListResponse, SavedViewResponse } from '../schemas/responses/views.js';
 import { createViewSchema, patchViewSchema, viewItemParamsSchema } from '../schemas/views.js';
 import {
